+++ raiden-dapp/src/views/Transfer.vue
@@ -94,11 +94,7 @@
       ></action-button>
     </v-container>
 
-    <error-dialog
-      :description="error"
-      :title="errorTitle"
-      @dismiss="error = ''"
-    ></error-dialog>
+    <error-dialog :error="error" @dismiss="error = null"></error-dialog>
   </v-form>
 </template>
 
@@ -117,7 +113,7 @@ import ChannelDepositDialog from '@/components/ChannelDepositDialog.vue';
 import DownArrow from '@/components/icons/DownArrow.vue';
 import { BigNumber } from 'ethers/utils';
 import { mapGetters, mapState } from 'vuex';
-import { RaidenChannel, ChannelState } from 'raiden-ts';
+import { RaidenChannel, ChannelState, RaidenError } from 'raiden-ts';
 import { Zero } from 'ethers/constants';
 import AddressUtils from '@/utils/address-utils';
 import NavigationMixin from '@/mixins/navigation-mixin';
@@ -154,8 +150,7 @@ export default class Transfer extends Mixins(BlockieMixin, NavigationMixin) {
   done: boolean = false;
   depositing: boolean = false;
 
-  errorTitle: string = '';
-  error: string = '';
+  error: Error | RaidenError | null = null;
 
   channels!: (tokenAddress: string) => RaidenChannel[];
 
@@ -206,7 +201,6 @@ export default class Transfer extends Mixins(BlockieMixin, NavigationMixin) {
 
   async deposit(amount: BigNumber) {
     this.loading = true;
-    this.errorTitle = this.$t('transfer.error.deposit-title') as string;
 
     try {
       await this.$raiden.deposit(
@@ -218,7 +212,7 @@ export default class Transfer extends Mixins(BlockieMixin, NavigationMixin) {
       this.loading = false;
       this.dismissProgress();
     } catch (e) {
-      this.error = e.message;
+      this.error = e;
       this.loading = false;
       this.depositing = false;
     }