+++ raiden-dapp/tests/unit/views/open-channel.spec.ts
@@ -11,10 +11,6 @@ import OpenChannel from '@/views/OpenChannel.vue';
 import Vue from 'vue';
 import Vuetify from 'vuetify';
 import { TestData } from '../data/mock-data';
-import RaidenService, {
-  ChannelDepositFailed,
-  ChannelOpenFailed
-} from '@/services/raiden-service';
 import store from '@/store/index';
 import NavigationMixin from '@/mixins/navigation-mixin';
 import { RouteNames } from '@/router/route-names';
@@ -23,6 +19,8 @@ import { Token } from '@/model/types';
 import { Tokens } from '@/types';
 import { mockInput } from '../utils/interaction-utils';
 import { parseUnits } from 'ethers/utils';
+import { RaidenError, ErrorCodes } from 'raiden-ts';
+import RaidenService from '@/services/raiden-service';
 
 Vue.use(Vuetify);
 Vue.filter('truncate', Filters.truncate);
@@ -53,7 +51,8 @@ describe('OpenChannel.vue', () => {
         $raiden: service,
         $router: router,
         $route: TestData.mockRoute(routeParams),
-        $t: (msg: string) => msg
+        $t: (msg: string) => msg,
+        $te: (msg: string) => msg
       }
     };
 
@@ -108,7 +107,7 @@ describe('OpenChannel.vue', () => {
 
     test('show an error when a channel open fails', async () => {
       service.openChannel.mockRejectedValueOnce(
-        new ChannelOpenFailed('open: transaction failed')
+        new RaidenError(ErrorCodes.CNL_OPENCHANNEL_FAILED)
       );
 
       mockInput(wrapper, '0.1');
@@ -117,20 +116,24 @@ describe('OpenChannel.vue', () => {
       button.trigger('click');
       await wrapper.vm.$nextTick();
       await flushPromises();
-      expect(wrapper.vm.$data.error).toBe('open-channel.error.open-failed');
+      expect(wrapper.vm.$data.error).toMatchObject({
+        code: 'CNL_OPENCHANNEL_FAILED'
+      });
       await flushPromises();
     });
 
     test('show an error when the deposit fails', async () => {
       service.openChannel.mockRejectedValueOnce(
-        new ChannelDepositFailed('deposit: transaction failed')
+        new RaidenError(ErrorCodes.RDN_DEPOSIT_TRANSACTION_FAILED)
       );
 
       mockInput(wrapper, '0.1');
       button.trigger('click');
       await wrapper.vm.$nextTick();
       await flushPromises();
-      expect(wrapper.vm.$data.error).toBe('open-channel.error.deposit-failed');
+      expect(wrapper.vm.$data.error).toMatchObject({
+        code: 'RDN_DEPOSIT_TRANSACTION_FAILED'
+      });
       await flushPromises();
     });
 
@@ -140,7 +143,7 @@ describe('OpenChannel.vue', () => {
       button.trigger('click');
       await wrapper.vm.$nextTick();
       await flushPromises();
-      expect(wrapper.vm.$data.error).toBe('unknown');
+      expect(wrapper.vm.$data.error).toMatchObject({ message: 'unknown' });
       await flushPromises();
     });