+++ raiden-ts/src/raiden.ts
@@ -62,7 +62,7 @@ import {
   chooseOnchainAccount,
   getContractWithSigner,
 } from './helpers';
-import RaidenError, { ErrorCodes } from './utils/error';
+import { RaidenError, ErrorCodes } from './utils/error';
 
 export class Raiden {
   private readonly store: Store<RaidenState, RaidenAction>;